% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_model_utilities.R
\name{dzip}
\alias{dzip}
\title{ZIP model pdf}
\usage{
dzip(x, lambda, pi)
}
\arguments{
\item{x}{Value the pdf is to be evaluated at}

\item{lambda}{The expected value for the Poisson model}

\item{pi}{The probability of being a structural zero}
}
\value{
The probability
}
\description{
Evaluates zip model at particular value k.
}
