% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_combine.R
\name{rf_combine}
\alias{rf_combine}
\title{Combine multiple regression trees}
\usage{
rf_combine(...)
}
\arguments{
\item{...}{A collection of random forest trees}
}
\value{
A random forest object
}
\description{
Combine multiple regression trees
}
