% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_doublecv.R
\name{rf_doublecv}
\alias{rf_doublecv}
\title{Fit zero-inflated random forests using double cross-validation.
In other words, nested cross-validation is used to estimate the
mse of a cross-validated zero-inflated random forest model.}
\usage{
rf_doublecv(x, z, y, ntree = 100, nfolds = 2, n_innerfolds = 4,
  tuning_mat = NULL, OuterFolds = NULL)
}
\arguments{
\item{x}{Matrix or data.frame of covariates for the count part
of the model.}

\item{z}{Matrix or data.frame of covariates for the zero-inflation
part of the model.}

\item{y}{vector of count outcomes.}

\item{ntree}{number of trees}

\item{nfolds}{number of folds used in outer cv}

\item{n_innerfolds}{number of folds used in inner cv rounds}

\item{tuning_mat}{matrix with tuning parameters to carry out cross-validation over}

\item{OuterFolds}{Index of which observation is in which outer cv fold.
This is important because it easily allows us to compare predictions from
multiple runs zero-inflated random forests or regular random forests.}
}
\value{
A list
}
\description{
Fit zero-inflated random forests using double cross-validation.
In other words, nested cross-validation is used to estimate the
mse of a cross-validated zero-inflated random forest model.
}
