% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_model_utilities.R
\name{XZ_to_form}
\alias{XZ_to_form}
\title{Produce formula from X and Z}
\usage{
XZ_to_form(out_name, X, Z)
}
\arguments{
\item{out_name}{outcome name}

\item{X}{design matrix for count model, should be a data.frame}

\item{Z}{design matrix for zero-inflation model, should be a data.frame}
}
\value{
A formula
}
\description{
The function takes data.frames X and Z, and it produces
a formula that could be used by other functions such
as pscl::zeroinfl.
}
