% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zirf_doublecv.R
\name{zirf_doublecv}
\alias{zirf_doublecv}
\title{Fit zero-inflated random forests using double cross-validation.
In other words, nested cross-validation is used to estimate the
mse of a cross-validated zero-inflated random forest model.}
\usage{
zirf_doublecv(x, z, y, rounds = rounds, nlambda = 20, ntree = 100,
  count_coef = NULL, zero_coef = NULL, nfolds = 2,
  n_innerfolds = 4, tuning_mat = NULL, OuterFolds = NULL,
  iter_keep_preds = NULL, iter_keep_pi = NULL, iter_keep_xsi = NULL)
}
\arguments{
\item{x}{Matrix or data.frame of covariates for the count part
of the model.}

\item{z}{Matrix or data.frame of covariates for the zero-inflation
part of the model.}

\item{y}{vector of count outcomes.}

\item{rounds}{how many iterations the EM algorithm should be run.}

\item{nlambda}{number folds used in estimating initial zip model}

\item{ntree}{number of trees}

\item{count_coef}{values for initial estimate of count model.}

\item{zero_coef}{values for initial estimate of zero-inflation model.}

\item{nfolds}{number of folds used in outer cv}

\item{n_innerfolds}{number of folds used in inner cv rounds}

\item{tuning_mat}{matrix with tuning parameters to carry out cross-validation over}

\item{OuterFolds}{Index of which observation is in which outer cv fold.
This is important because it easily allows us to compare predictions from
multiple runs zero-inflated random forests or regular random forests.}

\item{iter_keep_preds}{boolean, if TRUE prediction values are kept across EM iterations}

\item{iter_keep_pi}{boolean, if TRUE values of pi are kept across EM iterations}

\item{iter_keep_xsi}{boolean, if TRUE zero-inflation model coefficients
are kept across EM iterations}
}
\value{
A list
}
\description{
Fit zero-inflated random forests using double cross-validation.
In other words, nested cross-validation is used to estimate the
mse of a cross-validated zero-inflated random forest model.
}
