% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zirf_fit_func.R
\name{zirf_fit}
\alias{zirf_fit}
\title{Fit zero-inflated random forests given a pre-specified set of
tuning parameters.}
\usage{
zirf_fit(
  x,
  z,
  y,
  rounds = 25,
  mtry,
  ntree = 100,
  nodesize = 5,
  count_coef = NULL,
  zero_coef = NULL,
  newx = NULL,
  newz = NULL,
  iter_keep_pi = T,
  iter_keep_preds = T,
  iter_keep_xsi = T,
  nlambda = 10,
  threshold_prob = T,
  epsilon = 0.01
)
}
\arguments{
\item{x}{Matrix or data.frame of covariates for the count part
of the model.}

\item{z}{Matrix or data.frame of covariates for the zero-inflation
part of the model.}

\item{y}{vector of count outcomes.}

\item{rounds}{how many iterations the EM algorithm should be run at maximum.}

\item{mtry}{which value of mtry to use in each forest.}

\item{ntree}{number of trees.}

\item{nodesize}{controls depth of regression trees.}

\item{count_coef}{values for initial estimate of count model.}

\item{zero_coef}{values for initial estimate of zero-inflation model.}

\item{newx}{A list of count model covariate matrices for new data.}

\item{newz}{A list of zero inflation model covariate matrices for new data.}

\item{iter_keep_pi}{boolean, if TRUE values of pi are kept across EM iterations.}

\item{iter_keep_preds}{boolean, if TRUE prediction values are kept across EM iterations.}

\item{iter_keep_xsi}{boolean, if TRUE zero-inflation model coefficients
are kept across EM iterations.}

\item{nlambda}{number of folds of cross-validation.}

\item{threshold_prob}{boolean, if TRUE.}

\item{epsilon}{tolerance used for stopping iterations.
The maximum difference in probability of inclusion.
between iterations is the criterion.}
}
\value{
A list.
}
\description{
Fit zero-inflated random forests given a pre-specified set of
tuning parameters.
}
